tweaked mlGeNN TF module structure
