+++ ml_genn_tf/ml_genn_tf/converters/__init__.py
@@ -1,7 +1,10 @@
 from .converter import Converter
-from .enum import ConverterType, InputType
+from .enum import InputType
 from .simple import Simple
 from .few_spike import FewSpike
 from .data_norm import DataNorm
 
 from .spike_norm import spike_normalise
+
+__all__ = ["Converter", "DataNorm",
+           "FewSpike", "InputType", "Simple"]
