+++ ml_genn_tf/ml_genn_tf/converters/enum.py
@@ -4,9 +4,3 @@ class InputType(Enum):
     SPIKE = 'spike'
     POISSON = 'poisson'
     IF = 'if'
-
-class ConverterType(Enum):
-    SIMPLE = 'simple'
-    DATA_NORM = 'data-norm'
-    SPIKE_NORM = 'spike-norm'
-    FEW_SPIKE = 'few-spike'
